move TaggedElement type to layout module
since uses are constrained to this code
