+++ read-fonts/src/lib.rs
@@ -95,7 +95,7 @@ pub use offset::{Offset, ResolveNullableOffset, ResolveOffset};
 pub use offset_array::{ArrayOfNullableOffsets, ArrayOfOffsets};
 pub use read::{ComputeSize, FontRead, FontReadWithArgs, ReadArgs, ReadError, VarSize};
 pub use table_provider::{TableProvider, TopLevelTable};
-pub use table_ref::{TableRef, TaggedElement};
+pub use table_ref::TableRef;
 
 /// Public re-export of the font-types crate.
 pub extern crate font_types as types;
@@ -112,7 +112,7 @@ pub(crate) mod codegen_prelude {
         ComputeSize, FontRead, FontReadWithArgs, Format, ReadArgs, ReadError, VarSize,
     };
     pub use crate::table_provider::TopLevelTable;
-    pub use crate::table_ref::{TableRef, TaggedElement};
+    pub use crate::table_ref::TableRef;
     pub use std::ops::Range;
 
     pub use types::*;
